requirements.txt
PyYAML==6.0
pip-autoremove==0.10.0
orgparse==0.3.1
isort==5.10.1
flake8==4.0.1
black==22.6.0
